requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.8.0
python-dotenv==1.0.0
google-genai==0.8.0
openai==1.51.0
anthropic==0.36.0
langchain-core==0.3.0
langchain-google-genai==2.0.0